Hooking everything back up

+++ Data/Migrations/201604130012232_ClearOldHangfireJobs.cs
@@ -7,18 +7,40 @@ namespace Data.Migrations
         public override void Up()
         {
             Sql(@"
-DROP TABLE [HangFire].[AggregatedCounter]
-DROP TABLE [HangFire].[Counter]
-DROP TABLE [HangFire].[Hash]
-DROP TABLE [HangFire].[JobParameter]
-DROP TABLE [HangFire].[JobQueue]
-DROP TABLE [HangFire].[List]
-DROP TABLE [HangFire].[Schema]
-DROP TABLE [HangFire].[Server]
-DROP TABLE [HangFire].[Set]
-DROP TABLE [HangFire].[State]
-DROP TABLE [HangFire].[Job]
-TRUNCATE TABLE [dbo].[QueuedQuestionQueries]
+IF OBJECT_ID('*AggregatedCounter*', 'U') IS NOT NULL 
+    DROP TABLE [HangFire].[AggregatedCounter]
+
+IF OBJECT_ID('*Counter*', 'U') IS NOT NULL 
+    DROP TABLE [HangFire].[Counter]
+
+IF OBJECT_ID('*Hash*', 'U') IS NOT NULL 
+    DROP TABLE [HangFire].[Hash]
+
+IF OBJECT_ID('*JobParameter*', 'U') IS NOT NULL 
+    DROP TABLE [HangFire].[JobParameter]
+
+IF OBJECT_ID('*JobQueue*', 'U') IS NOT NULL 
+    DROP TABLE [HangFire].[JobQueue]
+
+IF OBJECT_ID('*List*', 'U') IS NOT NULL 
+    DROP TABLE [HangFire].[List]
+
+IF OBJECT_ID('*Schema*', 'U') IS NOT NULL 
+    DROP TABLE [HangFire].[Schema]
+
+IF OBJECT_ID('*Server*', 'U') IS NOT NULL 
+    DROP TABLE [HangFire].[Server]
+
+IF OBJECT_ID('*Set*', 'U') IS NOT NULL 
+    DROP TABLE [HangFire].[Set]
+
+IF OBJECT_ID('*State*', 'U') IS NOT NULL 
+    DROP TABLE [HangFire].[State]
+
+IF OBJECT_ID('*Job*', 'U') IS NOT NULL 
+    DROP TABLE [HangFire].[Job]
+
+DROP TABLE [dbo].[QueuedQuestionQueries]
 ");
         }
